Update AndroidManifest.xml files to omit package attribute and add explanatory comments
This commit modifies the AndroidManifest.xml files in both the network analysis and topology modules to omit the package attribute, aligning with the practices of the modern Android Gradle Plugin (AGP 7.0+). Additionally, comments have been added to clarify this change and direct developers to the relevant namespace definitions in the build.gradle.kts files.

Furthermore, a new "Reset" button has been introduced in the TopologyScreen, enhancing user interaction by allowing users to reset the state of the screen.